Initial Work for Project 1
Created remote database (on AWS). Created Login table and sample users.
Create Simple index.html page and allow user to submit login
information.
Create DOA and Service functions to support the login functionality.
The Login() servlet is the current service layer. It calls the DAO.
The DAO connects to the remote database using hardcoded credentials
(not included in the commit) and retrieves and prints all data in the
database. This is just test functionality for now.

Also commit supporting documents, including project notes and database
schema.

After this commit, Projet1 will be developed on a child-branch from
the 'West' branch.